Editor: Rename editorMediaUpload to mediaUpload
