+++ editor/utils/index.js
@@ -1,1 +1,20 @@
-export { default as editorMediaUpload } from './editor-media-upload';
+/**
+ * WordPress dependencies
+ */
+import deprecated from '@wordpress/deprecated';
+
+/**
+ * Internal dependencies
+ */
+import mediaUpload from './media-upload';
+
+export { mediaUpload };
+
+export function editorMediaUpload( ...params ) {
+	deprecated( 'wp.editor.editorMediaUpload', {
+		version: '3.5',
+		alternative: 'wp.editor.mediaUpload',
+		plugin: 'Gutenberg',
+	} );
+	mediaUpload( ...params );
+}
